Integration

+++ src/main/java/Hypercell/BlogApp/controller/CommentController.java
@@ -1,13 +1,15 @@
 package Hypercell.BlogApp.controller;
 
+import Hypercell.BlogApp.exceptions.GeneralException;
 import Hypercell.BlogApp.model.Comment;
+import Hypercell.BlogApp.model.response.body.GeneralResponse;
 import Hypercell.BlogApp.service.CommentService;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/comment")
+@RequestMapping("/api")
 public class CommentController {
     private final CommentService commentService;
 
@@ -16,25 +18,21 @@ public class CommentController {
         this.commentService = commentService;
     }
 
-    @GetMapping()
-    public List<Comment> getComments(){
 
-        return commentService.getComments();
+//    @GetMapping("/get/{comment-id}")
+//    public Comment getComment(@PathVariable("comment-id") Integer commentId) throws GeneralException {
+//        return commentService.getComment(commentId);
+//    }
+
+    @GetMapping("/all-comment/{post-id}")
+    public GeneralResponse<List<Comment>> getCommentByPost(@PathVariable("post-id") Integer postId) throws GeneralException {
+        return new GeneralResponse<>(true,commentService.getCommentByPost(postId));
     }
 
-    @GetMapping("/get/{comment-id}")
-    public Comment getComment(@PathVariable("comment-id") Integer commentId){
-        return commentService.getComment(commentId);
-    }
-
-    @GetMapping("/get/post/{post-id}")
-    public List<Comment> getCommentByPost(@PathVariable("post-id") Integer postId){
-        return commentService.getCommentByPost(postId);
-    }
-
-    @PostMapping("/add")
-    public Comment addComment(@RequestBody Comment comment){
-        return commentService.addComment(comment);
+    @PostMapping("/add-comment/{post-id}")
+    public GeneralResponse<Comment> addComment(@RequestBody Comment comment, @PathVariable("post-id") Integer postId){
+        comment.setPostId(postId);
+        return new GeneralResponse<>(true,commentService.addComment(comment));
     }
 
 
@@ -44,9 +42,14 @@ public class CommentController {
     }
 
 
-    @DeleteMapping("/delete/{comment-id}")
-    public void deleteComment(@PathVariable("comment-id") Integer id){
+    @DeleteMapping("/delete-comment/{comment-id}/{user-id}")
+    public GeneralResponse<Boolean> deleteComment(@PathVariable("comment-id") Integer id, @PathVariable("user-id") Integer userId) throws GeneralException {
+        if(commentService.getComment(id)==null)
+            throw new GeneralException("1","Wrong comment id entered");
+        if(commentService.getComment(id).getUser().getId()!=userId)
+            throw new GeneralException("1","You are not allowed to delete this comment");
         commentService.deleteComment(id);
+        return new GeneralResponse<>(true,true);
     }
